[WAYANG-#8] Change in names for readibility
