+++ python/src/pywy/graph/types.py
