+++ python/src/pywy/platforms/python/operator/__init__.py
@@ -1,0 +1,11 @@
+from pywy.platforms.python.operator.py_execution_operator import PyExecutionOperator
+from pywy.platforms.python.operator.py_unary_filter import PyFilterOperator
+from pywy.platforms.python.operator.py_source_textfile import PyTextFileSourceOperator
+from pywy.platforms.python.operator.py_sink_textfile import PyTextFileSinkOperator
+
+__ALL__ = [
+    PyExecutionOperator,
+    PyFilterOperator,
+    PyTextFileSourceOperator,
+    PyTextFileSinkOperator
+]
