+++ python/src/pywy/platforms/python/operator/py_source_textfile.py
@@ -1,6 +1,6 @@
 from typing import Set
-from pywy.operators.sink import TextFileSink
-from pywy.platforms.python.operators.PyExecutionOperator import PyExecutionOperator
+from pywy.operators.source import TextFileSource
+from pywy.platforms.python.operator.py_execution_operator import PyExecutionOperator
 from pywy.platforms.python.channels import (
                                                 Channel,
                                                 ChannelDescriptor,
@@ -9,29 +9,30 @@ from pywy.platforms.python.channels import (
                                             )
 
 
-class PyTextFileSinkOperator(TextFileSink, PyExecutionOperator):
+class PyTextFileSourceOperator(TextFileSource, PyExecutionOperator):
 
-    def __init__(self, origin: TextFileSink = None):
+    def __init__(self, origin: TextFileSource = None):
         path = None if origin is None else origin.path
         super().__init__(path)
         pass
 
     def execute(self, inputs: Channel, outputs: Channel):
         self.validate_channels(inputs, outputs)
-        if isinstance(inputs[0], PyIteratorChannel) :
-            file = open(self.path,'w')
-            py_in_iter_channel: PyIteratorChannel = inputs[0]
-            iterable = py_in_iter_channel.provide_iterable();
-            for element in iterable:
-                file.write(str(element))
-            file.close()
+        if isinstance(outputs[0], PyIteratorChannel) :
+            py_out_iter_channel: PyIteratorChannel = outputs[0]
+            py_out_iter_channel.accept_iterable(
+                open(
+                    self.path,
+                    'r'
+                )
+            )
 
         else:
             raise Exception("Channel Type does not supported")
 
 
     def get_input_channeldescriptors(self) -> Set[ChannelDescriptor]:
-        return {PyIteratorChannelDescriptor}
+        raise Exception("The PyTextFileSource does not support Input Channels")
 
     def get_output_channeldescriptors(self) -> Set[ChannelDescriptor]:
-        raise Exception("The PyTextFileSource does not support Output Channels")
+        return {PyIteratorChannelDescriptor}
