--- python/src/pywy/platforms/python/operators/__init__.py
@@ -1,11 +1,0 @@
-from pywy.platforms.python.operators.PyExecutionOperator import PyExecutionOperator
-from pywy.platforms.python.operators.PyFilterOperator import PyFilterOperator
-from pywy.platforms.python.operators.PyTextFileSourceOperator import PyTextFileSourceOperator
-from pywy.platforms.python.operators.PyTextFileSinkOperator import PyTextFileSinkOperator
-
-__ALL__ = [
-    PyExecutionOperator,
-    PyFilterOperator,
-    PyTextFileSourceOperator,
-    PyTextFileSinkOperator
-]
