--- python/src/pywy/platforms/python/platform/__init__.py
@@ -1,1 +1,0 @@
-from pywy.platforms.python.platform.platform import PythonPlatform
