+++ python/src/pywy/platforms/python/plugin.py
@@ -1,5 +1,5 @@
 from pywy.core import Executor
-from pywy.platforms.python.execution.executor import PyExecutor
+from pywy.platforms.python.execution import PyExecutor
 from pywy.platforms.python.platform import PythonPlatform
 from pywy.core import Plugin
 from pywy.platforms.python.mappings import PywyOperatorMappings
